fix(routes): block unpublished routes visit
