+++ src/views/auth/EmailConfirmationView.vue
@@ -94,6 +94,7 @@ import type {
   TutorStateInterface,
 } from "@/state/UserState.types";
 import { fillUserState, redirectByUserState } from "@/state/UserState";
+import router from '@/router';
 
 export default {
   name: "EmailConfirmationView",
@@ -117,15 +118,16 @@ export default {
         code: "",
       },
       authResolver: new AuthResolver(),
-      registrationData: (
-        JSON.parse(localStorage.getItem("dataToVerify") as string) as RegistrationData<
-          UserWithChildRegistrationDto | UserRegistrationDto,
-          TutorStateInterface | MentorStateInterface | ParentStateInterface
-        >
-      ).dto as UserRegistrationDto | UserWithChildRegistrationDto,
+      registrationData: null as null | UserWithChildRegistrationDto | UserRegistrationDto,
     };
   },
   mounted() {
+    this.registrationData = localStorage.getItem("dataToVerify")
+      ? (JSON.parse(localStorage.getItem("dataToVerify") as string) as RegistrationData<
+          UserWithChildRegistrationDto | UserRegistrationDto,
+          TutorStateInterface | MentorStateInterface | ParentStateInterface
+        >).dto as UserRegistrationDto | UserWithChildRegistrationDto
+      : null
     // Получаем email из query параметров или localStorage
     this.userEmail =
       this.$route.query.email as string ||
@@ -215,7 +217,7 @@ export default {
     },
 
     goBack() {
-      this.$router.push("/register");
+      router.go(-1)
     },
   },
 };
